Revert "SFS: cast hell"
This reverts commit 5b2bd909b7364859bcc4ca2c5ca669586be2b90f.

+++ arch/x86/kernel/sfi.c
@@ -98,8 +98,8 @@ int __init sfi_init_memory_map(void)
 	/* walk through the syst to search the mmap table */
 	mmapt = NULL;
 	for (i = 0; i < tbl_cnt; i++) {
-		if (!strncmp(SFI_SIG_MMAP, (char *)*pentry, 4)) {
-			mmapt = (struct sfi_table_simple *)*pentry;
+		if (!strncmp(SFI_SIG_MMAP, (char *)(u32)*pentry, 4)) {
+			mmapt = (struct sfi_table_simple *)(u32)*pentry;
 			break;
 		}
 		pentry++;
